카오스 던전 예상 수익 계산 (기존 API에 추가) - 코드 지저분해서 리팩토링 필요

+++ src/main/java/lostark/todo/domain/content/Content.java
@@ -19,6 +19,6 @@ public abstract class Content {
 
     private String name;
 
-    private int level; //컨텐츠 레벨
+    private double level; //컨텐츠 레벨
 
 }
